refactor(plugins): rename askUserPlugin → ask-user, align to template
Bring the oldest plugin into the canonical shape defined by
plugins/_template:

- folder: askUserPlugin/ → ask-user/ (kebab, no Plugin suffix)
- package: @hachej/boring-workspace-ask-user → @hachej/boring-ask-user
  (one naming convention: boring-<plugin-name>, treating it as a
  top-level plugin rather than a workspace sub-package)
- layout: flat front/server/shared/ → src/front, src/server, src/shared
- package.json: add `.` root export, `files: [dist]`, `sideEffects:
  false`, `peerDependencies` for workspace + react, devDep on
  workspace, @vitejs/plugin-react in devDeps
- tsup.config: nested entries (front/index → dist/front/index.js)
  and regex external `/^@hachej\/boring-/` to match the template
- vitest.config: drop cross-package borrow of
  packages/workspace/vitest.setup.ts in favor of
  plugins/_shared/vitest.setup.ts; add @vitejs/plugin-react and
  globals:true; add /events alias
- tsconfig: include `src/**/*`, add /events path alias

Consumers updated in apps/workspace-playground:
  package.json (dep + 4 build script filters), playwright.config.ts:7
  (e2e glob), src/front/App.tsx:4 (import), src/server/dev.ts:4
  (import). Source code under src/ unchanged — only the surrounding
  package shell moved.

Tests: 56 pass in ask-user, playground typecheck clean, lint:invariants
passes (272 files scanned).
