--- plugins/askUserPlugin/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@hachej/boring-workspace-ask-user",
-  "version": "0.1.13",
-  "type": "module",
-  "private": true,
-  "exports": {
-    "./front": {
-      "types": "./dist/front.d.ts",
-      "import": "./dist/front.js"
-    },
-    "./server": {
-      "types": "./dist/server.d.ts",
-      "import": "./dist/server.js"
-    },
-    "./shared": {
-      "types": "./dist/shared.d.ts",
-      "import": "./dist/shared.js"
-    }
-  },
-  "scripts": {
-    "build": "pnpm --filter @hachej/boring-ui-kit build && pnpm --filter @hachej/boring-workspace build && tsup",
-    "typecheck": "pnpm --filter @hachej/boring-ui-kit build && pnpm --filter @hachej/boring-workspace build && tsc --noEmit -p tsconfig.json",
-    "test": "pnpm --filter @hachej/boring-ui-kit build && vitest run"
-  },
-  "dependencies": {
-    "@hachej/boring-ui-kit": "workspace:*",
-    "@hachej/boring-workspace": "workspace:*",
-    "fastify": "^5.3.3",
-    "lucide-react": "^1.8.0",
-    "react": "^19.0.0",
-    "zod": "^3.23.0"
-  },
-  "devDependencies": {
-    "@testing-library/jest-dom": "^6.9.1",
-    "@testing-library/react": "^16.3.0",
-    "@types/node": "^22.0.0",
-    "jsdom": "^29.0.2",
-    "tsup": "^8.4.0",
-    "typescript": "~5.9.3",
-    "vitest": "^3.1.3"
-  }
-}
